+++ langgraph_agent_toolkit/agents/blueprints/chatbot/agent.py
@@ -1,12 +1,11 @@
 from langchain_core.messages import BaseMessage
 from langchain_core.runnables import RunnableConfig
-from langgraph.checkpoint.memory import MemorySaver
 from langgraph.func import entrypoint
-from langgraph.graph import add_messages
 
 from langgraph_agent_toolkit.agents.agent import Agent
 from langgraph_agent_toolkit.core import settings
 from langgraph_agent_toolkit.core.models.factory import ModelFactory
+from langgraph_agent_toolkit.schema.models import ModelProvider
 
 
 @entrypoint(
@@ -22,7 +21,12 @@ async def chatbot(
     if previous:
         messages = previous["messages"] + messages
 
-    model = ModelFactory.create(config["configurable"].get("model", settings.DEFAULT_MODEL_TYPE))
+    model = ModelFactory.create(
+        model_provider=ModelProvider.OPENAI,
+        model_name=settings.OPENAI_MODEL_NAME,
+        openai_api_base=settings.OPENAI_API_BASE_URL,
+        openai_api_key=settings.OPENAI_API_KEY,
+    )
     response = await model.ainvoke(messages)
     return entrypoint.final(value={"messages": [response]}, save={"messages": messages + [response]})